+++ .gitignore
@@ -1,1 +1,2 @@
 test/spec.browserify.js
+node_modules/
