+++ test/test.html
@@ -12,7 +12,7 @@
       mocha.ui('bdd');
       assert = chai.assert;
     </script>
-    <script src="../index.js"></script>
+    <script src="../jxon.js"></script>
     <script src="spec.js"></script>
     <script>
       mocha.run();
